On to part 5

+++ part4/bloglist/backend/controllers/blogs.js
@@ -2,7 +2,10 @@ const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("author", {
+    username: 1,
+    name: 1,
+  });
   response.status(200).json(blogs);
 });
 
@@ -13,23 +16,50 @@ blogsRouter.get("/:id", async (request, response) => {
 });
 
 blogsRouter.post("/", async (request, response) => {
-  const blog = new Blog(request.body);
-  const result = await blog.save();
+  const user = request.user;
+
+  if (!request.body.title || !request.body.url) {
+    return response.status(400).json({
+      error: "Blog title and url are required",
+    });
+  }
+
+  const blog = {
+    title: request.body.title,
+    url: request.body.url,
+    author: user.id,
+  };
+
+  const result = await Blog.create(blog);
+  user.blogs = user.blogs.concat(result.id);
+  await user.save();
   response.status(201).json(result);
 });
 
 blogsRouter.put("/:id", async (request, response) => {
+  const user = request.user;
   const id = request.params.id;
-  const result = await Blog.findByIdAndUpdate(id, request.body);
-  
-  response.status(200).json(result);
+  const blog = await Blog.findById(id);
+  if (blog.author.toString() === user.id.toString()) {
+    await Blog.findByIdAndUpdate(id, request.body);
+    response.status(200).json(blog);
+  } else {
+    response
+      .status(401)
+      .json({ error: "Blogs can only be edited by their respective author " });
+  }
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
   const id = request.params.id;
-
-  await Blog.findByIdAndDelete(id);
-  response.status(204).end();
+  const user = request.user;
+  const blog = await Blog.findById(id);
+  if (blog.author.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: "Blog can only be deleted by author " });
+  }
 });
 
 module.exports = blogsRouter;
